ADD: Implement waiting time after movement
The data from cameras lag behind the robot movement, hence wait
for a constant time to avoid basing the movement decision on
outdated data. Failure to do say causes overshooting when aiming
to correct for a robot pose that has an offset on a single
axis (say, the x-axis) compared to the desired pose.