MOD: Clean up robot interface, move logic to movement algorithm
Two major changes:

- Divide previous tune_robot method into two separate methods, one
moving along all axes at the same time, and the other one moving along
a single axis at a time. This is done because to support multiple
robots, it is important that the exact movement that methods at
the interface perform are well-defined, e.g., that they do not depend on
the robot. Previously, tune_robot made the robot move along all axes
if the robot supported it and along one axis at a time otherwise.

- Move consecutive axis movement logic from the robot class to the movement
algorithm; some algorithms may want to perform movements in a different
order or using a different logic altogether, hence the algorithm is the
correct place for that logic.