FIX: Detecting when robot has stopped moving
With newer, Linux-based Elfin, issuing a movement command
does not immediately result in robot returning "moving" as its
state. Instead, there is a short delay between issuing the command
and the robot reporting that it's moving.

To avoid interpreting that delay as robot having stopped moving, implement
a distinct state where we have issued the movement command but it has
not started yet.